backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pybaseball>=2.2.7
pandas>=2.2.3
numpy>=2.1.0
scikit-learn>=1.6.0
requests>=2.32.0
MLB-StatsAPI>=1.7.2
aiosqlite>=0.20.0
httpx>=0.27.0
python-dotenv>=1.0.1
